llm/requirements.txt
pandas==2.3.0
trino==0.336.0
pyarrow==20.0.0
deltalake==0.25.5
s3fs>=2023.1.0
requests>=2.28.0
python-dotenv>=1.0.1
langchain==0.3.27
langchain-openai==0.3.31
openai==1.101.0
